removed ts and updated the package to the latest version
