+++ packages/components/package.json
@@ -32,14 +32,14 @@
     "@babel/preset-react": "^7.16.7",
     "@storybook/addon-actions": "^6.5.10",
     "@storybook/addon-essentials": "^6.5.10",
+    "@storybook/addon-info": "^5.3.21",
     "@storybook/addon-interactions": "^6.5.10",
+    "@storybook/addon-knobs": "^6.4.0",
     "@storybook/addon-links": "^6.5.10",
     "@storybook/builder-webpack5": "^6.5.10",
     "@storybook/manager-webpack5": "^6.5.10",
     "@storybook/react": "^6.5.10",
     "@storybook/testing-library": "^0.0.13",
-    "@storybook/addon-info": "^5.3.21",
-    "@storybook/addon-knobs": "^6.4.0",
     "@testing-library/react": "^12.0.0",
     "@types/react": "^18.0.7",
     "@types/react-dom": "^18.0.0",
@@ -78,7 +78,7 @@
     "lodash.throttle": "^4.1.1",
     "prop-types": "^15.7.2",
     "react-content-loader": "^4.3.2",
-    "react-div-100vh": "^0.3.8",
+    "react-div-100vh": "^0.7.0",
     "react-dom": "^16.14.0",
     "react-drag-drawer": "^3.3.4",
     "react-dropzone": "11.0.1",
